Introduced IbexaKernelTestTrait to reduce code duplication
